requirements.txt
fastapi
sqlmodel
uvicorn
pytest
httpx
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
Python-JWT